Conflicts:
	mythtv/libs/libmythtv/tv_rec.cpp
	mythtv/programs/mythbackend/mainserver.cpp

+++ mythtv/programs/mythfrontend/galleryviewhelper.cpp
@@ -102,10 +102,6 @@ void GalleryViewHelper::LoadTreeData()
     QList<ImageMetadata *> *dirList = new QList<ImageMetadata *>;
     QList<ImageMetadata *> *fileList = new QList<ImageMetadata *>;
 
-    // Stop generating thumbnails
-    // when a new directory is loaded
-    m_fileHelper->StopThumbGen();
-
     // The parent id is the database index of the
     // directories which subdirectories and files shall be loaded
     int id = 0;
@@ -128,7 +124,7 @@ void GalleryViewHelper::LoadTreeData()
     // Clear the list so that it can be populated with new data.
     m_currentNode->deleteAllChildren();
 
-    // If the parent is a root directory
+    // If the node has a parent (ie. not the top level)
     // then add a additional directory at the beginning of the list that
     // is of the type kUpDirectory so that the user can navigate one level up.
     if (id > 0)
@@ -143,9 +139,6 @@ void GalleryViewHelper::LoadTreeData()
     // Load all files with the specified sorting criterias
     m_dbHelper->LoadFiles(fileList, id);
     LoadTreeNodeData(fileList, m_currentNode);
-
-    // Start generating thumbnails if required
-    m_fileHelper->StartThumbGen();
 
     // clean up
     delete dirList;
@@ -169,7 +162,8 @@ void GalleryViewHelper::LoadTreeNodeData(QList<ImageMetadata *> *list,
         ImageMetadata *im = list->at(i);
         if (im)
         {
-            m_fileHelper->AddToThumbnailList(im);
+            // don't recreate any existing thumbnails
+            m_fileHelper->AddToThumbnailList(im, false);
 
             // Create a new tree node that will hold the data
             MythGenericTree *treeItem =
@@ -191,45 +185,20 @@ void GalleryViewHelper::LoadTreeNodeData(QList<ImageMetadata *> *list,
  */
 void GalleryViewHelper::RenameCurrentNode(QString &newName)
 {
-    bool fileExist = false;
+    ImageMetadata *im = GetImageMetadataFromSelectedNode();
+    if (!im)
+        return;
 
-    // Check if the file with the new name already exists in the current directory
-    QList<MythGenericTree *> *nodeTree = m_currentNode->getAllChildren();
-    for (int i = 0; i < nodeTree->size(); i++)
+    if (m_fileHelper->RenameFile(im, newName))
     {
-        ImageMetadata *im = nodeTree->at(i)->GetData().value<ImageMetadata *>();
-        if (im)
-        {
-            if (im->m_name.compare(newName) == 0)
-            {
-                fileExist = true;
-                break;
-            }
-        }
-    }
+        // replace the original filename with the
+        // new one in the pull path + filename variable
+        QString newFileName = im->m_fileName.replace(im->m_name, newName);
 
-    // The file with the given new name does not yet exist.
-    // Continue with the renaming of the file
-    if (!fileExist)
-    {
-        ImageMetadata *im = GetImageMetadataFromSelectedNode();
-        if (!im)
-            return;
-
-        if (m_fileHelper->RenameFile(im, newName))
-        {
-            // replace the original filename with the
-            // new one in the pull path + filename variable
-            QString newFileName = im->m_fileName.replace(im->m_name, newName);
-
-            im->m_fileName = newFileName;
-            im->m_name = newName;
-
-            m_dbHelper->UpdateData(im);
-        }
+        im->m_fileName = newFileName;
+        im->m_name = newName;
     }
 }
-
 
 
 /** \fn     GalleryViewHelper::DeleteCurrentNode()
@@ -239,24 +208,20 @@ void GalleryViewHelper::RenameCurrentNode(QString &newName)
 void GalleryViewHelper::DeleteCurrentNode()
 {
     ImageMetadata *im = GetImageMetadataFromSelectedNode();
-    if (!im)
-        return;
 
-    // Delete the file and remove the database entry
-    if (m_fileHelper->RemoveFile(im))
-    {
-        // Remove the entry from the node list
-        m_currentNode->deleteNode(m_currentNode->getSelectedChild());
+    // TODO: Remove directories as well
+    if (im && im->m_type >= kImageFile)
 
-        // Only remove the first thumbnail when it is an image or video.
-        // If its a folder, it will use thumbnails from other images
-        if (im->m_type == kImageFile || im->m_type == kVideoFile)
-            QFile::remove(im->m_thumbFileNameList->at(0));
+        // Delete the file and remove the database entry
+        if (m_fileHelper->RemoveFile(im))
+        {
+            // Remove the entry from the node list
+            m_currentNode->deleteNode(m_currentNode->getSelectedChild());
 
-        m_dbHelper->RemoveFile(im);
-    }
+            // Clean up thumbnail cache
+            GetMythUI()->RemoveFromCacheByFile(im->m_thumbFileNameList->at(0));
+        }
 }
-
 
 
 /** \fn     GalleryViewHelper::DeleteSelectedNodes()
@@ -269,18 +234,15 @@ void GalleryViewHelper::DeleteSelectedNodes()
     for (int i = 0; i < nodeTree->size(); i++)
     {
         ImageMetadata *im = nodeTree->at(i)->GetData().value<ImageMetadata *>();
-        if (im && im->m_selected)
+
+        // TODO: Remove directories as well
+        if (im && im->m_selected && im->m_type >= kImageFile)
         {
-            // Delete the file and remove the
-            // database entry only if it was successful
+            // Delete the file and remove the database entry
             if (m_fileHelper->RemoveFile(im))
             {
-                // Only remove the first thumbnail when it is an image or video.
-                // If its a folder, it will use thumbnails from other images
-                if (im->m_type == kImageFile || im->m_type == kVideoFile)
-                    QFile::remove(im->m_thumbFileNameList->at(0));
-
-                m_dbHelper->RemoveFile(im);
+                // Clean up thumbnail cache
+                GetMythUI()->RemoveFromCacheByFile(im->m_thumbFileNameList->at(0));
 
                 // Remove the entry from the node list
                 m_currentNode->deleteNode(nodeTree->at(i));
@@ -372,7 +334,6 @@ void GalleryViewHelper::SetNodeVisibilityState(int nodeState)
 }
 
 
-
 /** \fn     GalleryViewHelper::SetFileOrientation(int)
  *  \brief  Saves the orientation information of the selected node
  *  \param  fileOrientation The orientation value 1-8
@@ -391,25 +352,21 @@ void GalleryViewHelper::SetFileOrientation(int fileOrientation)
     // value will then be saved in the exif header tag.
     im->SetOrientation(fileOrientation, false);
 
-    // Update the exif tag, if that fails we can restore the original
-    // orientation so that the database and image file are not out of sync
+    // Request orientation update
     if (m_fileHelper->SetImageOrientation(im))
     {
-        m_dbHelper->UpdateData(im);
-        m_fileHelper->RecreateThumbnail(im);
-        m_fileHelper->StartThumbGen();
+        // force thumbnail to be regenerated
+        m_fileHelper->AddToThumbnailList(im, true);
     }
     else
     {
-        im->SetOrientation(oldFileOrientation, false);
+        // Restore previous orientation
+        im->SetOrientation(oldFileOrientation, true);
 
-        LOG(VB_GENERAL, LOG_ERR,
-            QString("Could not write the angle %1 into the file %2. "
-                    "The database value has not been updated.")
-            .arg(im->GetAngle()).arg(im->m_fileName));
+        LOG(VB_GENERAL, LOG_ERR, QString("Orientation update failed for %1")
+            .arg(im->m_fileName));
     }
 }
-
 
 
 /** \fn     GalleryViewHelper::SetFileZoom(int)
@@ -424,10 +381,10 @@ void GalleryViewHelper::SetFileZoom(int zoom)
         return;
 
     if (zoom == kFileZoomIn)
-        im->SetZoom(20);
+        im->SetZoom(20, false);
 
     if (zoom == kFileZoomOut)
-        im->SetZoom(-20);
+        im->SetZoom(-20, false);
 
     m_dbHelper->UpdateData(im);
 }
@@ -498,22 +455,6 @@ ImageMetadata *GalleryViewHelper::GetImageMetadataFromNode(int i)
 
 
 
-/** \fn     GalleryViewHelper::SetPreviewImageSize(MythUIButtonList *)
- *  \brief  Gets and saves the size of the shown image previews which
- *          will be used to create thumbnails of the given dimensions.
- *  \param  imageList The widget where the themer has specified the images sizes
- *  \return void
- */
-void GalleryViewHelper::SetPreviewImageSize(MythUIButtonList *imageList)
-{
-    float width  = (float)imageList->ItemWidth();
-    float height = (float)imageList->ItemHeight();
-
-    m_fileHelper->SetThumbnailSize((int)width, (int)height);
-}
-
-
-
 /** \fn     GallerySyncStatusThread::GallerySyncStatusThread()
  *  \brief  Constructor
  *  \return void
@@ -547,37 +488,42 @@ bool GallerySyncStatusThread::isSyncRunning()
  */
 void GallerySyncStatusThread::run()
 {
-    volatile bool exit = false;
+    bool syncDetected = false;
+
     GalleryFileHelper *fh = new GalleryFileHelper();
 
     // Internal counter that tracks how many
     // times we have been in the while loop
     int loopCounter = 0;
 
-    while (!exit)
+    while (true)
     {
         GallerySyncStatus status = fh->GetSyncStatus();
 
         LOG(VB_GENERAL, LOG_DEBUG,
-            QString("GallerySyncStatusThread: Sync status is running: %1, Syncing image '%2' of '%3'")
+            QString("GallerySyncStatusThread: Sync status is running: %1, "
+                    "Syncing image '%2' of '%3'")
             .arg(status.running).arg(status.current).arg(status.total));
 
-        // Only update the progress text
-        // if the sync is still running
+        // Update the progress text whilst the sync is running
         if (status.running)
-            emit UpdateSyncProgress(status.current, status.total);
-
-        // Try at least one time to get the sync
-        // status before checking for the exit condition
-        if (loopCounter >= 1)
         {
-            if (status.running == false)
-                exit = true;
+            syncDetected = true;
+            emit UpdateSyncProgress(status.current, status.total);
         }
+        // Check at least twice before quitting
+        else if (loopCounter >= 1)
+        {
+            // only refresh UI after syncs
+            if (syncDetected)
+                emit SyncComplete();
 
+            // die
+            break;
+        }
         // Wait some time before trying to get and update the status
         // This also avoids too many calls to the service api.
-        usleep(999999);
+        usleep(500000);
 
         ++loopCounter;
     }
